+++ cloudalibaba-provider-rbac9003/src/main/java/com/bosssoft/RBAC/PO/RoleMenu.java
@@ -4,6 +4,7 @@ import lombok.Data;
 
 import javax.persistence.Column;
 import javax.persistence.Id;
+import java.math.BigInteger;
 import java.time.LocalDateTime;
 
 /**
@@ -14,11 +15,11 @@ import java.time.LocalDateTime;
 @Data
 public class RoleMenu {
     @Id
-    private Integer id;
+    private BigInteger id;
     @Column
-    private Integer roleId;
+    private BigInteger roleId;
     @Column
-    private Integer menuId;
+    private BigInteger menuId;
     @Column
     private LocalDateTime createTime;
     @Column
